[add] altered ci with PR comments. made test helper methods more modular
