+++ tests/include.py
@@ -1,0 +1,60 @@
+import glob
+import os
+
+MEMTIER_BINARY = os.environ.get("MEMTIER_BINARY", "memtier_benchmark")
+
+
+def assert_minimum_memtier_outcomes(config, env, memtier_ok, merged_command_stats, overall_expected_request_count,
+                                    overall_request_count):
+    # assert correct exit code
+    env.assertTrue(memtier_ok == True)
+    # assert we have all outputs
+    env.assertTrue(os.path.isfile('{0}/mb.stdout'.format(config.results_dir)))
+    env.assertTrue(os.path.isfile('{0}/mb.stderr'.format(config.results_dir)))
+    env.assertTrue(os.path.isfile('{0}/mb.json'.format(config.results_dir)))
+
+    # assert we have the expected request count
+    env.assertEqual(overall_expected_request_count, overall_request_count)
+
+
+def add_required_env_arguments(benchmark_specs, config, env, master_nodes_list):
+    # check if environment is cluster
+    if env.isCluster():
+        benchmark_specs["args"].append("--cluster-mode")
+    # check if environment uses Unix Socket connections
+    if env.isUnixSocket():
+        benchmark_specs["args"].append("--unix-socket")
+        benchmark_specs["args"].append(master_nodes_list[0]['unix_socket_path'])
+        config["memtier_benchmark"]['explicit_connect_args'] = True
+    else:
+        config['redis_process_port'] = master_nodes_list[0]['port']
+
+
+def get_expected_request_count(config):
+    result = -1
+    if 'memtier_benchmark' in config:
+        mt = config['memtier_benchmark']
+        if 'threads' in mt and 'clients' in mt and 'requests' in mt:
+            result = config['memtier_benchmark']['threads'] * config['memtier_benchmark']['clients'] * \
+                     config['memtier_benchmark']['requests']
+    return result
+
+
+def agg_info_commandstats(master_nodes_connections, merged_command_stats):
+    overall_request_count = 0
+    for master_connection in master_nodes_connections:
+        shard_stats = master_connection.execute_command("INFO", "COMMANDSTATS")
+        for cmd_name, cmd_stat in shard_stats.items():
+            if cmd_name in merged_command_stats:
+                overall_request_count += cmd_stat['calls']
+                merged_command_stats[cmd_name]['calls'] = merged_command_stats[cmd_name]['calls'] + cmd_stat['calls']
+    return overall_request_count
+
+
+def ensure_clean_benchmark_folder(dirname):
+    files = glob.glob('{}/*'.format(dirname))
+    for f in files:
+        os.remove(f)
+    if os.path.exists(dirname):
+        os.removedirs(dirname)
+    os.makedirs(dirname)
