+++ tests/test_requirements.txt
@@ -1,0 +1,2 @@
+git+https://github.com/RedisLabsModules/RLTest.git@master
+git+https://github.com/RedisLabs/mbdirector.git@master
